refactored backstack initialization
