+++ change/@microsoft-teams-js-557e4bd6-ceda-49b6-9936-17ed2ce73916.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Refactored backstack initialization to resolve an issue with dynamically importing `pages`",
+  "packageName": "@microsoft/teams-js",
+  "email": "109628470+noahdarveau-MSFT@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
