+++ packages/teams-js/src/internal/handlers.ts
@@ -3,10 +3,11 @@
 import { ApiName, ApiVersionNumber, getApiVersionTag } from '../internal/telemetry';
 import { FrameContexts } from '../public/constants';
 import { HostToAppPerformanceMetrics, LoadContext, ResumeContext } from '../public/interfaces';
-import * as pages from '../public/pages/pages';
+//import * as pages from '../public/pages/pages';
 import { runtime } from '../public/runtime';
 import { Communication, sendMessageEventToChild, sendMessageToParent } from './communication';
 import { ensureInitialized } from './internalAPIs';
+import { handleBackButtonPress, pagesTelemetryVersionNumber } from './pagesHelpers';
 import { getLogger } from './telemetry';
 import { isNullOrUndefined } from './typeCheckUtilities';
 
@@ -43,7 +44,13 @@ class HandlersPrivate {
     HandlersPrivate.handlers['themeChange'] = handleThemeChange;
     HandlersPrivate.handlers['load'] = handleLoad;
     HandlersPrivate.handlers['beforeUnload'] = handleBeforeUnload;
-    pages.backStack._initialize();
+    //pages.backStack._initialize();
+    registerHandler(
+      getApiVersionTag(pagesTelemetryVersionNumber, ApiName.Pages_BackStack_RegisterBackButtonPressHandler),
+      'backButtonPress',
+      handleBackButtonPress,
+      false,
+    );
   }
 
   /**
